Add collision detection to Class Living_being
Clean up in human and zombie files, and add default collision detection.
In order for collision detection to work, the object must be initialized
with a group of sprites containing the sprites that the collision
detection would work against.

Collision detection can be tuned for each sub class by reimplementing the
valid_movement() function in each class.

Also I add a process_collisions() method to the base class and it gets
called from the update method. This is to check the other collisions like
power ups, weapons and terrain objects.